Update benchmarks and simplify color numeric bulk colormatrix
