+++ tests/ImageSharp.Benchmarks/Bulk/Rgb24Bytes.cs
@@ -3,14 +3,13 @@
 
 using System.Buffers;
 using BenchmarkDotNet.Attributes;
-
 using SixLabors.ImageSharp.Memory;
 using SixLabors.ImageSharp.PixelFormats;
 
 // ReSharper disable InconsistentNaming
-namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk;
+namespace SixLabors.ImageSharp.Benchmarks.Bulk;
 
-public abstract class ToRgba32Bytes<TPixel>
+public abstract class Rgb24Bytes<TPixel>
     where TPixel : unmanaged, IPixel<TPixel>
 {
     private IMemoryOwner<TPixel> source;
@@ -27,7 +26,7 @@ public abstract class ToRgba32Bytes<TPixel>
     {
         this.configuration = Configuration.Default;
         this.source = this.configuration.MemoryAllocator.Allocate<TPixel>(this.Count);
-        this.destination = this.configuration.MemoryAllocator.Allocate<byte>(this.Count * 4);
+        this.destination = this.configuration.MemoryAllocator.Allocate<byte>(this.Count * 3);
     }
 
     [GlobalCleanup]
@@ -37,28 +36,9 @@ public abstract class ToRgba32Bytes<TPixel>
         this.destination.Dispose();
     }
 
-    // [Benchmark]
-    public void Naive()
-    {
-        Span<TPixel> s = this.source.GetSpan();
-        Span<byte> d = this.destination.GetSpan();
-
-        for (int i = 0; i < this.Count; i++)
-        {
-            TPixel c = s[i];
-            int i4 = i * 4;
-            Rgba32 rgba = default;
-            c.ToRgba32(ref rgba);
-            d[i4] = rgba.R;
-            d[i4 + 1] = rgba.G;
-            d[i4 + 2] = rgba.B;
-            d[i4 + 3] = rgba.A;
-        }
-    }
-
     [Benchmark(Baseline = true)]
     public void CommonBulk() =>
-        new PixelOperations<TPixel>().ToRgba32Bytes(
+        new PixelOperations<TPixel>().ToRgb24Bytes(
             this.configuration,
             this.source.GetSpan(),
             this.destination.GetSpan(),
@@ -66,21 +46,13 @@ public abstract class ToRgba32Bytes<TPixel>
 
     [Benchmark]
     public void OptimizedBulk() =>
-        PixelOperations<TPixel>.Instance.ToRgba32Bytes(
+        PixelOperations<TPixel>.Instance.ToRgb24Bytes(
             this.configuration,
             this.source.GetSpan(),
             this.destination.GetSpan(),
             this.Count);
 }
 
-public class ToRgba32Bytes_FromRgba32 : ToRgba32Bytes<Rgba32>
-{
-}
-
-public class ToRgba32Bytes_FromArgb32 : ToRgba32Bytes<Argb32>
-{
-}
-
-public class ToRgba32Bytes_FromBgra32 : ToRgba32Bytes<Bgra32>
+public class Rgb24Bytes_Rgba32 : Rgb24Bytes<Rgba32>
 {
 }
