+++ tests/ImageSharp.Benchmarks/General/GetSetPixel.cs
@@ -12,7 +12,7 @@ public class GetSetPixel
     [Benchmark(Baseline = true, Description = "System.Drawing GetSet pixel")]
     public System.Drawing.Color GetSetSystemDrawing()
     {
-        using var source = new Bitmap(400, 400);
+        using Bitmap source = new(400, 400);
         source.SetPixel(200, 200, System.Drawing.Color.White);
         return source.GetPixel(200, 200);
     }
@@ -20,7 +20,7 @@ public class GetSetPixel
     [Benchmark(Description = "ImageSharp GetSet pixel")]
     public Rgba32 GetSetImageSharp()
     {
-        using var image = new Image<Rgba32>(400, 400);
+        using Image<Rgba32> image = new(400, 400);
         image[200, 200] = Color.White;
         return image[200, 200];
     }
